// types.ts
export interface ProductFAQ {
  question: string;
  answer: string;
}

// Table: market_shibari
export interface Product {
  id: number;
  title: string;
  title_en?: string;
  color: string;
  color_en?: string;
  description_short: string;
  description_short_en?: string;
  description_long: string;
  description_long_en?: string;
  price?: number;
  images: string[]; // URLs array
  video_url?: string;
  faq?: ProductFAQ[];
  created_at?: string;
  is_course?: boolean; // Helper flag to distinguish in cart
  category?: string; // New field for dynamic categories
}
// ...
export interface CartItem {
  product: Product;
  quantity: number;
}